[LB-393] refactor : 채점 및 레포트 코드 리팩토링
