+++ src/main/java/com/lgcms/leveltest/service/mamberanswer/MemberAnswerUpdateService.java
@@ -1,4 +1,4 @@
-package com.lgcms.leveltest.service;
+package com.lgcms.leveltest.service.mamberanswer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
